+++ sdk/evaluation/azure-ai-evaluation/tests/e2etests/test_evaluate.py
@@ -13,10 +13,12 @@ from azure.ai.evaluation import (
     ContentSafetyEvaluator,
     ContentSafetyMultimodalEvaluator,
     SexualMultimodalEvaluator,
+    ProtectedMaterialMultimodalEvaluator,
     F1ScoreEvaluator,
     FluencyEvaluator,
     GroundednessEvaluator,
     GroundednessProEvaluator,
+    RetrievalEvaluator,
     evaluate,
 )
 from azure.ai.evaluation._common.math import list_mean_nan_safe
@@ -45,6 +47,12 @@ def data_convo_file():
 def multimodal_file_with_imageurls():
     data_path = os.path.join(pathlib.Path(__file__).parent.resolve(), "data")
     return os.path.join(data_path, "dataset_messages_image_urls.jsonl")
+
+
+@pytest.fixture
+def multimodal_file_with_imageurls_with_target():
+    data_path = os.path.join(pathlib.Path(__file__).parent.resolve(), "data")
+    return os.path.join(data_path, "dataset_messages_image_urls_target.jsonl")
 
 
 @pytest.fixture
@@ -251,7 +259,7 @@ class TestEvaluate:
             azure_ai_project=project_scope, credential=azure_cred, _parallel=False
         )
         result = evaluate(
-            evaluation_name=f"test-mm-eval-dataset-img-url-{str(uuid.uuid4())}",
+            evaluation_name=f"test-mm-content-safety-eval-dataset-img-url-{str(uuid.uuid4())}",
             azure_ai_project=project_scope,
             data=multimodal_file_with_imageurls,
             evaluators={"content_safety": content_safety_eval},
@@ -283,23 +291,23 @@ class TestEvaluate:
         assert 0 <= metrics.get("content_safety.hate_unfairness_defect_rate") <= 1
 
     def test_evaluate_with_content_safety_multimodal_evaluator_with_target(
-        self, project_scope, azure_cred, multimodal_file_with_imageurls
+        self, project_scope, azure_cred, multimodal_file_with_imageurls_with_target
     ):
         os.environ["PF_EVALS_BATCH_USE_ASYNC"] = "false"
         from .target_fn import target_multimodal_fn1
 
-        input_data = pd.read_json(multimodal_file_with_imageurls, lines=True)
+        input_data = pd.read_json(multimodal_file_with_imageurls_with_target, lines=True)
         content_safety_eval = ContentSafetyMultimodalEvaluator(
             azure_ai_project=project_scope, credential=azure_cred, _parallel=False
         )
         result = evaluate(
             evaluation_name=f"test-mm-eval-dataset-img-url-target-{str(uuid.uuid4())}",
             azure_ai_project=project_scope,
-            data=multimodal_file_with_imageurls,
+            data=multimodal_file_with_imageurls_with_target,
             target=target_multimodal_fn1,
             evaluators={"content_safety": content_safety_eval},
             evaluator_config={
-                "content_safety": {"conversation": "${data.conversation}"},
+                "content_safety": {"conversation": "${target.conversation}"},
             },
         )
 
@@ -406,6 +414,44 @@ class TestEvaluate:
         # Check that label is renamed to passing rate in metrics
         assert "groundedness_pro.groundedness_pro_passing_rate" in convo_metrics.keys()
         assert 0 <= convo_metrics.get("groundedness_pro.groundedness_pro_passing_rate") <= 1
+
+    def test_evaluate_with_protected_material_multimodal_evaluator(
+        self, project_scope, azure_cred, multimodal_file_with_imageurls
+    ):
+        os.environ["PF_EVALS_BATCH_USE_ASYNC"] = "false"
+        input_data = pd.read_json(multimodal_file_with_imageurls, lines=True)
+        eval = ProtectedMaterialMultimodalEvaluator(azure_ai_project=project_scope, credential=azure_cred)
+        result = evaluate(
+            evaluation_name=f"test-multimodal-protected-material-eval-dataset-{str(uuid.uuid4())}",
+            azure_ai_project=project_scope,
+            data=multimodal_file_with_imageurls,
+            evaluators={"protected_material": eval},
+            evaluator_config={
+                "protected_material": {"conversation": "${data.conversation}"},
+            },
+        )
+
+        row_result_df = pd.DataFrame(result["rows"])
+        metrics = result["metrics"]
+        # validate the results
+        assert result is not None
+        assert result["rows"] is not None
+        assert row_result_df.shape[0] == len(input_data)
+
+        assert "outputs.protected_material.artwork_label" in row_result_df.columns.to_list()
+        assert "outputs.protected_material.artwork_reason" in row_result_df.columns.to_list()
+        assert "outputs.protected_material.fictional_characters_label" in row_result_df.columns.to_list()
+        assert "outputs.protected_material.fictional_characters_reason" in row_result_df.columns.to_list()
+        assert "outputs.protected_material.logos_and_brands_label" in row_result_df.columns.to_list()
+        assert "outputs.protected_material.logos_and_brands_reason" in row_result_df.columns.to_list()
+
+        assert "protected_material.fictional_characters_label" in metrics.keys()
+        assert "protected_material.logos_and_brands_label" in metrics.keys()
+        assert "protected_material.artwork_label" in metrics.keys()
+
+        assert 0 <= metrics.get("protected_material.fictional_characters_label") <= 1
+        assert 0 <= metrics.get("protected_material.logos_and_brands_label") <= 1
+        assert 0 <= metrics.get("protected_material.artwork_label") <= 1
 
     # @pytest.mark.performance_test
     @pytest.mark.skip(reason="Temporary skip to merge 37201, will re-enable in subsequent pr")
@@ -819,4 +865,63 @@ class TestEvaluate:
         assert row_result_df["outputs.grounded.groundedness"][1] in [3, 4, 5]
         assert row_result_df["outputs.grounded.evaluation_per_turn"][0]["groundedness"][0] in [3.0, 4.0, 5.0]
         assert row_result_df["outputs.grounded.evaluation_per_turn"][0]["groundedness_reason"][0] is not None
+        assert result["studio_url"] is None
+
+    def test_evaluate_with_retrieval_evaluator_with_convo(self, model_config, data_convo_file):
+        # data
+        input_data = pd.read_json(data_convo_file, lines=True)
+
+        retrieval_eval = RetrievalEvaluator(model_config)
+
+        # run the evaluation
+        result = evaluate(
+            data=data_convo_file,
+            evaluators={"retrieval": retrieval_eval},
+        )
+
+        row_result_df = pd.DataFrame(result["rows"])
+        metrics = result["metrics"]
+
+        # validate the results
+        assert result is not None
+        assert result["rows"] is not None
+        assert row_result_df.shape[0] == len(input_data)
+
+        assert "outputs.retrieval.retrieval" in row_result_df.columns.to_list()
+        assert "outputs.retrieval.gpt_retrieval" in row_result_df.columns.to_list()
+        assert "outputs.retrieval.evaluation_per_turn" in row_result_df.columns.to_list()
+        assert "retrieval.retrieval" in metrics.keys()
+        assert "retrieval.gpt_retrieval" in metrics.keys()
+        assert metrics.get("retrieval.retrieval") == list_mean_nan_safe(row_result_df["outputs.retrieval.retrieval"])
+        assert row_result_df["outputs.retrieval.retrieval"][1] in [0.0, 1.0, 2.0]
+        assert row_result_df["outputs.retrieval.evaluation_per_turn"][0]["retrieval"][0] in [0.0, 1.0, 2.0]
+        assert row_result_df["outputs.retrieval.evaluation_per_turn"][0]["retrieval_reason"][0] is not None
+        assert result["studio_url"] is None
+
+    def test_evaluate_with_retrieval_evaluator_with_column_mappings(self, model_config, data_file):
+        # data
+        input_data = pd.read_json(data_file, lines=True)
+
+        retrieval_eval = RetrievalEvaluator(model_config)
+
+        # run the evaluation
+        result = evaluate(
+            data=data_file,
+            evaluators={"retrieval": retrieval_eval},
+        )
+
+        row_result_df = pd.DataFrame(result["rows"])
+        metrics = result["metrics"]
+
+        # validate the results
+        assert result is not None
+        assert result["rows"] is not None
+        assert row_result_df.shape[0] == len(input_data)
+
+        assert "outputs.retrieval.retrieval" in row_result_df.columns.to_list()
+        assert "outputs.retrieval.gpt_retrieval" in row_result_df.columns.to_list()
+        assert "retrieval.retrieval" in metrics.keys()
+        assert "retrieval.gpt_retrieval" in metrics.keys()
+        assert metrics.get("retrieval.retrieval") == list_mean_nan_safe(row_result_df["outputs.retrieval.retrieval"])
+        assert row_result_df["outputs.retrieval.retrieval"][1] in [0.0, 1.0, 2.0]
         assert result["studio_url"] is None